chore: Move Sell landing and FAQ page to Sell app (#14413)
* chore: Move Sell landing page and faq page to Sell app

* comment
