+++ src/Apps/Sell/Routes/MarketingLanding/Components/LandingPage/SpeakToTheTeam.tsx
@@ -1,8 +1,8 @@
 import { ActionType, ContextModule } from "@artsy/cohesion"
 import { Box, Button, Image, Text } from "@artsy/palette"
-import { TextAndImageLayout } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/TextAndImageLayout"
-import { useAnalyticsContext } from "System/Hooks/useAnalyticsContext"
+import { TextAndImageLayout } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/TextAndImageLayout"
 import { RouterLink } from "System/Components/RouterLink"
+import { useAnalyticsContext } from "System/Hooks/useAnalyticsContext"
 import { useSystemContext } from "System/Hooks/useSystemContext"
 import { resized } from "Utils/resized"
 import { useTracking } from "react-tracking"
