+++ src/Apps/Sell/Routes/MarketingLanding/Components/__tests__/HowItWorksSteps.jest.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, render, screen } from "@testing-library/react"
-import { HowItWorksSteps } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/HowItWorksSteps"
+import { HowItWorksSteps } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/HowItWorksSteps"
 import { useTracking } from "react-tracking"
 import { useSystemContext } from "System/Hooks/useSystemContext"
