+++ src/Apps/Sell/Routes/MarketingLanding/MarketingLandingApp.tsx
@@ -1,17 +1,17 @@
 import { Box, Flex, Join, Spacer } from "@artsy/palette"
-import { CollectorsOverview } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/CollectorsOverview"
-import { FAQSWA } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/FAQSWA"
-import { FooterBanner } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/FooterBanner"
-import { HeaderSWA } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/HeaderSWA"
-import { Highlights } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/Highlights"
-import { HowItWorksSteps } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/HowItWorksSteps"
-import { MeetTheSpecialistsQueryRenderer } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/MeetTheSpecialists"
-import { PreviouslySoldOnArtsyRailQueryRenderer } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/PreviouslySoldOnArtsyRail"
-import { SWAFooter } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/SWAFooter"
-import { SWAStickyFooter } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/SWAStickyFooter"
-import { SpeakToTheTeam } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/SpeakToTheTeam"
-import { WaysWeSell } from "Apps/Consign/Routes/MarketingLanding/Components/LandingPage/WaysWeSell"
-import { SellMeta } from "Apps/Consign/Routes/MarketingLanding/Components/SellMeta"
+import { CollectorsOverview } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/CollectorsOverview"
+import { FAQSWA } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/FAQSWA"
+import { FooterBanner } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/FooterBanner"
+import { HeaderSWA } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/HeaderSWA"
+import { Highlights } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/Highlights"
+import { HowItWorksSteps } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/HowItWorksSteps"
+import { MeetTheSpecialistsQueryRenderer } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/MeetTheSpecialists"
+import { PreviouslySoldOnArtsyRailQueryRenderer } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/PreviouslySoldOnArtsyRail"
+import { SWAFooter } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/SWAFooter"
+import { SWAStickyFooter } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/SWAStickyFooter"
+import { SpeakToTheTeam } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/SpeakToTheTeam"
+import { WaysWeSell } from "Apps/Sell/Routes/MarketingLanding/Components/LandingPage/WaysWeSell"
+import { SellMeta } from "Apps/Sell/Routes/MarketingLanding/Components/SellMeta"
 import { Footer } from "Components/Footer/Footer"
 import { useRouter } from "System/Hooks/useRouter"
 import { Media } from "Utils/Responsive"
