+++ src/Apps/Sell/Routes/NewRoute.tsx
@@ -1,7 +1,7 @@
 import { FullBleed } from "@artsy/palette"
 import { AppContainer } from "Apps/Components/AppContainer"
-import { SellMeta } from "Apps/Consign/Routes/MarketingLanding/Components/SellMeta"
 import { ArtistRoute } from "Apps/Sell/Routes/ArtistRoute"
+import { SellMeta } from "Apps/Sell/Routes/MarketingLanding/Components/SellMeta"
 import { SellFlowContextProvider } from "Apps/Sell/SellFlowContext"
 import { EnableRecaptcha } from "Utils/EnableRecaptcha"
 import * as React from "react"
