+++ src/Apps/Sell/Routes/SubmissionRoute.tsx
@@ -1,6 +1,6 @@
 import { FullBleed } from "@artsy/palette"
 import { AppContainer } from "Apps/Components/AppContainer"
-import { SellMeta } from "Apps/Consign/Routes/MarketingLanding/Components/SellMeta"
+import { SellMeta } from "Apps/Sell/Routes/MarketingLanding/Components/SellMeta"
 import { SellFlowContextProvider } from "Apps/Sell/SellFlowContext"
 import { SubmissionRoute_submission$key } from "__generated__/SubmissionRoute_submission.graphql"
 import { HttpError } from "found"
